Moved to template and lib function
